+++ Data/ByteArray/Methods.hs
@@ -52,11 +52,6 @@ import           Foreign.Ptr
 import           Prelude hiding (length, take, drop, span, reverse, concat, replicate, splitAt, null, pred, last, any, all)
 import qualified Prelude
 
-#if defined(WITH_BYTESTRING_SUPPORT) && defined(WITH_BASEMENT_SUPPORT)
-import qualified Data.ByteString as SPE (ByteString)
-import qualified Basement.UArray as SPE (UArray)
-import qualified Basement.Block  as SPE (Block)
-#endif
 
 -- | Allocate a new bytearray of specific size, and run the initializer on this memory
 alloc :: ByteArray ba => Int -> (Ptr p -> IO ()) -> IO ba
@@ -304,9 +299,3 @@ all f b = not (any (not . f) b)
 -- | Convert a bytearray to another type of bytearray
 convert :: (ByteArrayAccess bin, ByteArray bout) => bin -> bout
 convert bs = inlineUnsafeCreate (length bs) (copyByteArrayToPtr bs)
-#if defined(WITH_BYTESTRING_SUPPORT) && defined(WITH_BASEMENT_SUPPORT)
-{-# SPECIALIZE convert :: SPE.ByteString -> SPE.UArray Word8 #-}
-{-# SPECIALIZE convert :: SPE.UArray Word8 -> SPE.ByteString #-}
-{-# SPECIALIZE convert :: SPE.ByteString -> SPE.Block Word8 #-}
-{-# SPECIALIZE convert :: SPE.Block Word8 -> SPE.ByteString #-}
-#endif