+++ Data/ByteArray/Types.hs
@@ -31,17 +31,6 @@ import           Data.Memory.PtrMethods (memCopy)
 import           Data.Proxy (Proxy(..))
 import           Data.Word (Word8)
 
-import qualified Basement.Types.OffsetSize as Base
-import qualified Basement.UArray as Base
-import qualified Basement.String as Base (String, toBytes, Encoding(UTF8))
-
-import qualified Basement.UArray.Mutable as BaseMutable (withMutablePtrHint)
-import qualified Basement.Block as Block
-import qualified Basement.Block.Mutable as Block
-
-import           Basement.Nat
-import qualified Basement.Sized.Block as BlockN
-
 import Prelude hiding (length)
 
 -- | Class to Access size properties and data of a ByteArray
@@ -75,59 +64,3 @@ instance ByteArray Bytestring.ByteString where
         return (r, Bytestring.PS fptr 0 sz)
 #endif
 
-#ifdef WITH_BASEMENT_SUPPORT
-
-baseBlockRecastW8 :: Base.PrimType ty => Block.Block ty -> Block.Block Word8
-baseBlockRecastW8 = Block.unsafeCast -- safe with Word8 destination
-
-instance Base.PrimType ty => ByteArrayAccess (Block.Block ty) where
-    length a = let Base.CountOf i = Block.length (baseBlockRecastW8 a) in i
-    withByteArray a f = Block.withPtr (baseBlockRecastW8 a) (f . castPtr)
-    copyByteArrayToPtr ba dst = do
-        mb <- Block.unsafeThaw (baseBlockRecastW8 ba)
-        Block.copyToPtr mb 0 (castPtr dst) (Block.length $ baseBlockRecastW8 ba)
-
-instance (KnownNat n, Base.PrimType ty, Base.Countable ty n) => ByteArrayAccess (BlockN.BlockN n ty) where
-    length a = let Base.CountOf i = BlockN.lengthBytes a in i
-    withByteArray a f = BlockN.withPtr a (f . castPtr)
-    copyByteArrayToPtr bna = copyByteArrayToPtr (BlockN.toBlock bna)
-
-baseUarrayRecastW8 :: Base.PrimType ty => Base.UArray ty -> Base.UArray Word8
-baseUarrayRecastW8 = Base.recast
-
-instance Base.PrimType ty => ByteArrayAccess (Base.UArray ty) where
-    length a = let Base.CountOf i = Base.length (baseUarrayRecastW8 a) in i
-    withByteArray a f = Base.withPtr (baseUarrayRecastW8 a) (f . castPtr)
-    copyByteArrayToPtr ba dst = Base.copyToPtr ba (castPtr dst)
-
-instance ByteArrayAccess Base.String where
-    length str = let Base.CountOf i = Base.length bytes in i
-      where
-        -- the Foundation's length return a number of elements not a number of
-        -- bytes. For @ByteArrayAccess@, because we are using an @Int@, we
-        -- didn't see that we were returning the wrong @CountOf@.
-        bytes = Base.toBytes Base.UTF8 str
-    withByteArray s f = withByteArray (Base.toBytes Base.UTF8 s) f
-
-instance (Ord ty, Base.PrimType ty) => ByteArray (Block.Block ty) where
-    allocRet sz f = do
-        mba <- Block.new $ sizeRecastBytes sz Proxy
-        a   <- Block.withMutablePtrHint True False mba (f . castPtr)
-        ba  <- Block.unsafeFreeze mba
-        return (a, ba)
-
-instance (Ord ty, Base.PrimType ty) => ByteArray (Base.UArray ty) where
-    allocRet sz f = do
-        mba <- Base.new $ sizeRecastBytes sz Proxy
-        a   <- BaseMutable.withMutablePtrHint True False mba (f . castPtr)
-        ba  <- Base.unsafeFreeze mba
-        return (a, ba)
-
-sizeRecastBytes :: Base.PrimType ty => Int -> Proxy ty -> Base.CountOf ty
-sizeRecastBytes w p = Base.CountOf $
-    let (q,r) = w `Prelude.quotRem` szTy
-     in q + (if r == 0 then 0 else 1)
-  where !(Base.CountOf szTy) = Base.primSizeInBytes p
-{-# INLINE [1] sizeRecastBytes #-}
-
-#endif